Add game pause, Improve Highscores
Improve Highscores - now there is a label in game over that tells you if you beat a highscore.

Pause - added game pausing. It still needs a pause panel though.
Also when game is paused the "time since start" isnt updated so it is as if it was still running. Needs a fix.